Add Grades query to GraphQL
The resolvers in this commit might just be the most poorly designed code I've ever written.
Essentially what we do is mock a call to Rest grades helper, then format that response to graphql.
It's horrible.

After HackUCI, we need to destroy all of the resolvers in this commit and restructure the project.
Helper functions should be shared between rest and graphql